requirements.txt
pyside2==5.14.2.2
pyserial
watchdog
pillow